migrate used code from external_rl to own code
