--- external_rl/.gitignore
@@ -1,3 +1,0 @@
-*__pycache__
-*egg-info
-.vscode
