--- external_rl/requirements.txt
@@ -1,5 +1,0 @@
-torch-ac>=1.4.0
-minigrid>=2.2.0
-tensorboardX>=1.6
-numpy>=1.3
-gymnasium>=0.26
