--- external_rl/utils/__init__.py
@@ -1,5 +1,0 @@
-from .agent import *
-from .env import *
-from .format import *
-from .other import *
-from .storage import *
