+++ src/ppo/utils/env.py
@@ -1,0 +1,7 @@
+import gymnasium as gym
+
+
+def make_env(env_key, seed=None, render_mode=None):
+    env = gym.make(env_key, render_mode=render_mode)
+    env.reset(seed=seed)
+    return env
