+++ src/ppo/utils/other.py
@@ -1,0 +1,29 @@
+import collections
+import random
+
+import numpy
+import torch
+
+
+device = torch.device(
+    "cuda"
+    if torch.cuda.is_available()
+    else ("mps" if torch.backends.mps.is_available() else "cpu")
+)
+
+
+def seed(seed):
+    random.seed(seed)
+    numpy.random.seed(seed)
+    torch.manual_seed(seed)
+    if torch.cuda.is_available():
+        torch.cuda.manual_seed_all(seed)
+
+
+def synthesize(array):
+    d = collections.OrderedDict()
+    d["mean"] = numpy.mean(array)
+    d["std"] = numpy.std(array)
+    d["min"] = numpy.amin(array)
+    d["max"] = numpy.amax(array)
+    return d
